# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Exame Perto — contrato de planejamento
  version: 0.4.0
  description: 'Autenticação e autorização familiar implementadas nos tickets 03–04; demais operações permanecem planejamento. Autorizar cada ação por papel e registro. Integrações
    ausentes retornam 503 INTEGRATION_UNAVAILABLE; política material indefinida retorna 422 POLICY_UNDEFINED. Erros
    principais: INVALID_INPUT, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, STATE_CONFLICT, FUNDING_INSUFFICIENT, IDEMPOTENCY_CONFLICT,
    ALREADY_ASSIGNED, VERSION_MISMATCH, RATE_LIMITED, LOCATION_STALE. Nomes de rotas em inglês e estados de domínio
    em português são estáveis nesta proposta. Respostas sensíveis usam Cache-Control: no-store. Listas sempre paginadas
    após escopo de autorização. Este arquivo não especifica STOMP; ver docs/SEGURANCA.md. Administração de papéis,
    políticas, instituições e habilitação de integrações depende de procedimentos ainda não definidos; nenhuma rota
    pública permite promovê-los. Replay de comando concluído exige sessão e relação histórica autorizadas; estado
    ativo é guarda só para execução inédita. Revogação de representação prevalece sobre replay. O retorno mínimo
    de replay não reabre acesso a GPS/endereço/código. D01–D12 adotadas em 30/09/2026: arquitetura não habilita
    políticas ou integrações. Cadastro/segurança e particular independem de programa subsidiado. AD/AO/AB/GF exigem
...
servers:
- url: /api/v1
...
security:
- bearerAuth: []
paths:
...
  /me/privacy-requests:
    get:
      operationId: listOwnPrivacyRequests
      summary: Listar solicitações próprias
      tags: [me]
      x-permission: Conta autenticada; somente protocolos próprios
      responses:
        '200': {description: Protocolos próprios, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/PrivacyRequest'}}}}}
        '401': {$ref: '#/components/responses/Error401'}
    post:
      operationId: requestPrivacy
      summary: Registrar solicitação do titular
      tags:
      - me
      description: Conta autenticada; demanda encaminhada para processo auditado. Prazo e tratamento dependem de
        política validada. Não promete exclusão imediata de obrigação legal.
      x-permission: Conta autenticada; demanda encaminhada para processo auditado
      responses:
        '202':
          description: Registrar solicitação do titular
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequest'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '503':
          $ref: '#/components/responses/Error503'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequestInput'
  /me/profile:
    get:
      operationId: getOwnProfile
      summary: Consultar perfil básico próprio
      tags: [me]
      description: Retorna somente os campos permitidos da própria conta. E-mail não é alterado por esta operação; conta encerrada perde a sessão e deve ser acompanhada pelo protocolo de privacidade.
      x-permission: Conta autenticada; próprio perfil
      responses:
        '200':
          description: Perfil próprio
          headers:
            ETag: {schema: {type: string}}
          content: {application/json: {schema: {$ref: '#/components/schemas/Profile'}}}
        '401': {$ref: '#/components/responses/Error401'}
        '404': {$ref: '#/components/responses/Error404'}
    put:
      operationId: updateOwnProfile
      summary: Editar campos permitidos do perfil próprio
      tags: [me]
      description: Atualiza nome e telefone com controle otimista de versão. Não altera e-mail, estado, identidade, políticas ou papéis.
      x-permission: Conta autenticada; próprio perfil; Origin/CSRF web
      parameters:
      - name: If-Match
        in: header
        required: true
        schema: {type: string, pattern: '^"[0-9]+"$'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProfileInput'}}}
      responses:
        '200': {description: Perfil atualizado, content: {application/json: {schema: {$ref: '#/components/schemas/Profile'}}}}
        '400': {$ref: '#/components/responses/Error400'}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '404': {$ref: '#/components/responses/Error404'}
        '412': {$ref: '#/components/responses/Error412'}
        '428': {$ref: '#/components/responses/Error428'}
  /me/privacy-requests/{id}:
    get:
      operationId: getPrivacyRequest
      summary: Acompanhar solicitação própria
      tags: [me]
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: Protocolo e estado, sem conteúdo de terceiros
          content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyRequest'}}}
        '401': {$ref: '#/components/responses/Error401'}
        '404': {$ref: '#/components/responses/Error404'}
  /privacy-requests/{id}/response:
    post:
      operationId: respondPrivacyRequest
      summary: Responder solicitação sem afirmar que eventual exclusão foi executada
      tags: [privacy]
      x-permission: Analista operacional nominal com MFA; autorização revalidada na chamada
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyResponseInput'}}}
      responses:
        '200': {description: "Solicitação respondida ou negada, ainda sem comprovar expurgo", content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyRequest'}}}}
        '400': {$ref: '#/components/responses/Error400'}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '404': {$ref: '#/components/responses/Error404'}
        '409': {$ref: '#/components/responses/Error409'}
  /privacy-requests/{id}/purge:
    post:
      operationId: authorizePrivacyPurge
      summary: Autorizar expurgo com política validada e inventário de recursos
      tags: [privacy]
      x-permission: Analista operacional nominal com MFA; configuração explícita de expurgo
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: "Expurgo autorizado, não executado", content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyRequest'}}}}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '404': {$ref: '#/components/responses/Error404'}
        '409': {$ref: '#/components/responses/Error409'}
        '422': {$ref: '#/components/responses/Error422'}
        '503': {$ref: '#/components/responses/Error503'}
  /privacy-requests/{id}/purge/execute:
    post:
      operationId: executePrivacyPurge
      summary: Executar ou retomar expurgo idempotente dos recursos controlados
      tags: [privacy]
      x-permission: Analista operacional nominal com MFA; papel e sessão revalidados
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: Estado da execução; pendências externas e preservações permanecem explícitas, content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyRequest'}}}}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
        '503': {$ref: '#/components/responses/Error503'}
  /privacy-requests/{id}/purge/verify:
    post:
      operationId: verifyPrivacyPurge
      summary: Verificar ausência nos recursos efetivamente controlados
      tags: [privacy]
      description: VERIFICADA limita-se aos recursos locais controlados. Dispositivos, fornecedores e backups continuam como procedimento pendente e não recebem comprovação fictícia.
      x-permission: Analista operacional nominal com MFA; papel e sessão revalidados
      parameters:
      - name: id
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: Verificação local concluída com contadores de ressalvas, content: {application/json: {schema: {$ref: '#/components/schemas/PrivacyRequest'}}}}
        '401': {$ref: '#/components/responses/Error401'}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
        '503': {$ref: '#/components/responses/Error503'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sessão revogável, emissor/audiência/assinatura/expiração validados. Nenhum privilégio inferido
        de cadastro.
    paymentSignature:
      type: apiKey
      in: header
      name: X-Event-Signature
      description: 'Placeholder técnico do adaptador: cabeçalho sozinho não autentica. Algoritmo, janela e headers
        finais precisam seguir o provedor selecionado; endpoint desabilitado até isso.'
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh
      description: Refresh rotativo da sessão WEB, sempre acompanhado de CSRF/Origin. MOBILE envia refresh no corpo.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
      description: Única por ator/operação. Mesmo corpo retorna resultado original após reautorização; corpo diferente
        ou em andamento retorna 409. Chaves financeiras de negócio não expiram.
...
  schemas:
...
    Version:
      type: integer
      format: int64
      minimum: 0
    Error:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
          format: uuid
        retryable:
          type: boolean
      required:
      - code
      - message
      - correlationId
      - retryable
      description: Sem dados pessoais, stack trace ou existência de recurso alheio. code é estável; message é segura
        para exibição.
...
    PrivacyRequestInput:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum:
          - ACESSO
          - CORRECAO
          - EXCLUSAO
        description:
          type: string
          minLength: 1
          maxLength: 4000
      required:
      - type
      - description
    ProfileInput:
      type: object
      additionalProperties: false
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        phone: {type: string, maxLength: 40, nullable: true}
      required: [name]
    Profile:
      type: object
      additionalProperties: false
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        name: {type: string}
        phone: {type: string, nullable: true}
        status: {type: string, enum: [ATIVO]}
        emailVerifiedAt: {type: string, format: date-time, nullable: true}
        version: {$ref: '#/components/schemas/Version'}
      required: [id, email, name, phone, status, emailVerifiedAt, version]
    PrivacyResponseInput:
      type: object
      additionalProperties: false
      required: [status, response]
      properties:
        status:
          type: string
          enum: [RESPONDIDA, NEGADA]
        response:
          type: string
          minLength: 1
          maxLength: 4000
...
    PrivacyRequest:
      type: object
      additionalProperties: false
      required:
      - id
      - protocol
      - type
      - status
      - createdAt
      - purgeStatus
      - procedurePending
      - preserved
      - closureStatus
      - anonymizationStatus
      - version
      properties:
        id:
          type: string
          format: uuid
        protocol:
          type: string
          description: Protocolo de acompanhamento; não comprova exclusão executada ou verificada.
        type:
          type: string
          enum:
          - ACESSO
          - CORRECAO
          - EXCLUSAO
        status:
          type: string
          enum:
          - PENDENTE
          - EM_ANALISE
          - RESPONDIDA
          - NEGADA
          - EXPURGO_SOLICITADO
          - EXPURGADA
          - VERIFICADA
        createdAt:
          type: string
          format: date-time
        response:
          type: string
          maxLength: 4000
          nullable: true
        purgeStatus:
          type: string
          nullable: true
          enum: [AUTORIZADA, EXECUTADA, VERIFICADA, FALHA]
          description: Estado operacional do expurgo; FALHA permanece retomável e não equivale a exclusão.
        executedAt:
          type: string
          format: date-time
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        procedurePending:
          type: integer
          minimum: 0
          description: Recursos fora do controle local que ainda dependem de procedimento ou contrato.
        preserved:
          type: integer
          minimum: 0
          description: Recursos preservados por regra explícita, incluindo finanças enquanto a retenção estiver pendente.
        closureStatus:
          type: string
          nullable: true
          enum: [NAO_APLICAVEL, PENDENTE, BLOQUEADA, EM_EXECUCAO, ENCERRADA, FALHA]
          description: Estado separado do fechamento da conta; BLOQUEADA indica custódia, obrigação ou reconciliação pendente.
        anonymizationStatus:
          type: string
          enum: [NAO_APLICAVEL, PENDENTE, TRANSFORMADA_NAO_INTEGRAL]
          description: Identificadores controlados foram transformados, mas mapeamentos internos e referências financeiras impedem declarar anonimização integral.
        version:
          $ref: '#/components/schemas/Version'
...
  responses:
    Error400:
      description: Requisição inválida / cursor inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error401:
      description: Autenticação ausente, expirada ou revogada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: Ação negada, CSRF ou escopo insuficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: Recurso inexistente ou não visível ao ator.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: Conflito de estado, saldo insuficiente, duplicidade ou idempotência incompatível.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error412:
      description: Versão If-Match obsoleta.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Error422:
      description: Regra não satisfeita, evidência inválida, política indefinida ou GPS antigo.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error428:
      description: If-Match obrigatório ausente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error429:
      description: Limite de tentativas/requisições excedido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Segundos até nova tentativa, quando conhecidos.
    Error503:
      description: Integração ou configuração operacional indisponível; nunca resposta fictícia.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
          description: Segundos até nova tentativa, quando conhecidos.